Added Audio key
Using the OfflineAudioContext API created a key for the fingerprint. The mobile support for the OfflineAudioContext is not great so it jumps over if mobile is detected.

Tested also the AudioContext API but does not work all the time. 

Code and info about Audio Fingerprint found here :https://audiofingerprint.openwpm.com/